--- app/src/main/java/kr/co/landvibe/handicraft/data/source/furniture/local/FurnitureLocalDataSource.java
@@ -1,14 +1,0 @@
-package kr.co.landvibe.handicraft.data.source.furniture.local;
-
-
-public class FurnitureLocalDataSource {
-    //싱글톤
-
-
-    int  foo(){
-        int a=2;
-
-        return a*a;
-    }
-
-}
